feat: block user delete untill access keys deleted (#144)

+++ internal/controller/user/user.go
@@ -244,6 +244,15 @@ func (c *external) Delete(ctx context.Context, mg resource.Managed) (managed.Ext
 		GroupID: cr.Spec.ForProvider.GroupID,
 		UserID:  meta.GetExternalName(mg),
 	}
+
+	creds, err := c.cloudianService.ListUserCredentials(ctx, user)
+	if err != nil {
+		return managed.ExternalDelete{}, err
+	}
+	if len(creds) > 0 {
+		return managed.ExternalDelete{}, errors.New("User has access keys and cannot be deleted")
+	}
+
 	if err := c.cloudianService.DeleteUser(ctx, user); err != nil {
 		return managed.ExternalDelete{}, errors.Wrap(err, errDeleteUser)
 	}
